+++ src/modules/securitybaseline/src/so/SecurityBaselineModule.c
@@ -2,8 +2,6 @@
 // Licensed under the MIT License.
 
 #include <Mmi.h>
-#include <sys/stat.h>
-#include <unistd.h>
 #include <SecurityBaseline.h>
 
 void __attribute__((constructor)) InitModule(void)
